+++ rust/src/04/4766.rs
@@ -2,7 +2,7 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let temperatures: Vec<f32> = buf
+    let temperatures: Vec<f64> = buf
         .lines()
         .take_while(|&s| s != "999")
         .flat_map(str::parse)
