+++ rust/src/24/24313.rs
@@ -2,7 +2,7 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let input = buf.split_ascii_whitespace().flat_map(str::parse::<f32>);
+    let input = buf.split_ascii_whitespace().flat_map(str::parse::<f64>);
     let [a1, a0, c, n0] = input.collect::<Vec<_>>()[..] else { return };
 
     if c == a1 {
